massive clean up and abstraction of game component and battle feild. player controlls cleaned up and abstracted out
