+++ src/components/UI/GameComponent/GameComponent.jsx
@@ -13,7 +13,7 @@ const GameComponent = (props) => {
     currentLevel: 1,
     maxFloor: 1,
     currentFloor: 1,
-    wave: state.computer.waveGenerator(state.sceneManager, state.difficulty),
+    wave: state.computer.waveGenerator(state.sceneManager[0], state.difficulty),
   });
   const [battle, setBattle] = useState({
     selectedCard: null,
