+++ src/components/UI/GameComponent/GameComponent_Battefield.jsx
@@ -1,78 +1,9 @@
 import { ClickableHeros, ClickableMonsters } from './GameComponent_Clickables';
 import { PartyStats, MonsterStats } from './GameComponent_Stats';
+import { BattleHero, BattleMonster } from './GameComponent_Battle_station';
 import { Link } from 'react-router-dom';
 import { PartyStatusBar, EnemyStatusBar } from './GameComponent_StatusBars';
-const VictoryComponent = (props) => {
-  const { state, setState } = props;
-
-  const handleLevelChange = () => {
-    setState({
-      ...state,
-      currentWave: 0,
-      floor: state.computer.waveGenerator(state.sceneManager, state.difficulty),
-    });
-    state.computer.isBossDefeated = false;
-  };
-  return (
-    <div>
-      <h1>You Completed The Floor</h1>
-      <Link to="/dashboard/story">
-        <button onClick={handleLevelChange}>
-          <h2>Next Floor</h2>
-        </button>
-      </Link>
-      <Link to="/dashboard/story">
-        <button onClick={handleLevelChange}>
-          <h2>Return To Town</h2>
-        </button>
-      </Link>
-    </div>
-  );
-};
-
-const VictoryMessageLootBox = (props) => {
-  const { state, setState } = props;
-
-  const victoryMessage = <VictoryComponent state={state} setState={setState} />;
-
-  return (
-    <div className="battlefield">
-      <div className="battle-animation">
-        {state.computer.isBossDefeated ? victoryMessage : <h1>VS</h1>}
-      </div>
-      <div id="LootBox">
-        {state.story[state.currentStory].isReward ? <h2>lootbox</h2> : null}
-      </div>
-    </div>
-  );
-};
-
-const BattleHero = (props) => {
-  const { battle, setBattle, state, setState } = props;
-
-  const handleTankClick = () => {
-    setBattle({
-      ...battle,
-      targetHero: state.party[0],
-    });
-  };
-  return (
-    <div onClick={handleTankClick} id="BattleHero">
-      <div id="monster-attack-animation">{battle.monsterDamageAnimation}</div>
-      <img src={state.tank.imgUrl} alt="" />
-    </div>
-  );
-};
-
-const BattleMonster = (props) => {
-  const { battle } = props;
-  return (
-    <div id="BattleMonster">
-      <div id="hero-attack-animation">{battle.heroDamageAnimation}</div>
-      <img src={battle.targetMonster.imgUrl} alt="" />
-    </div>
-  );
-};
+import { VictoryMessageLootBox } from './GameComponent_LootBox';
 
 const Battlefield = (props) => {
   const {
