+++ src/components/UI/GameComponent/GameComponent_Cards.jsx
@@ -1,0 +1,27 @@
+const PlayingCards = ({ setState, state }) => {
+  return (
+    <div className="playing-cards">
+      {state.playerHand.length ? (
+        <MappedPlayingCards state={state} setState={setState} />
+      ) : (
+        <h2>You Need Some Cards to Do Battle!</h2>
+      )}
+    </div>
+  );
+};
+const MappedPlayingCards = ({ state }) => {
+  return (
+    <div id="PlayingCards">
+      {state.playerHand.map((card) => {
+        return (
+          <div key={card.id}>
+            <h1>{card.name}</h1>
+            <img src={card.imgUrl} alt="" />
+          </div>
+        );
+      })}
+    </div>
+  );
+};
+
+export { PlayingCards };
