+++ src/components/UI/GameComponent/GameComponent_Stats.jsx
@@ -2,8 +2,8 @@ const PartyStats = (props) => {
   const { battle } = props;
   return (
     <div id="PartyStats">
-      <h1>Current Hero - {battle.targetHero.name}</h1>
-      <h2>Health -{battle.targetHero.health}</h2>
+      <h1>{battle.targetHero.name}</h1>
+      <h2>Health {battle.targetHero.health}</h2>
       <p>Attatck Power - {battle.targetHero.attackPower}</p>
     </div>
   );
@@ -13,8 +13,8 @@ const MonsterStats = (props) => {
   const { battle } = props;
   return (
     <div id="MonsterStats">
-      <h1>Current Target - {battle.targetMonster.name}</h1>
-      <h2>Health -{battle.targetMonster.health}</h2>
+      <h1>{battle.targetMonster.name}</h1>
+      <h2>Health {battle.targetMonster.health}</h2>
       <p>Attatck Power - {battle.targetMonster.attackPower}</p>
     </div>
   );
